backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pyftdi>=0.55.0
victron-ble>=0.4.0
httpx>=0.27.0
starlink-grpc-core>=0.0.28
